+++ modules/integrations/splunk_o11y_conf_shared/dashboards.tf
@@ -1,5 +1,5 @@
 resource "signalfx_dashboard_group" "forgecicd" {
-  name        = "ForgeCICD Dashboards"
+  name        = var.dashboard_group_name
   description = ""
   teams       = [var.team]
 
@@ -91,11 +91,24 @@ module "dashboard_forge_impact" {
     signalfx = signalfx
   }
 
-  tenant_names    = var.dashboard_variables.runner_k8s.tenant_names
-  dashboard_group = signalfx_dashboard_group.forgecicd.id
+  tenant_names      = try(var.dashboard_variables.forge_impact.tenant_names, var.dashboard_variables.runner_k8s.tenant_names)
+  dynamic_variables = try(var.dashboard_variables.forge_impact.dynamic_variables, [])
+  dashboard_group   = signalfx_dashboard_group.forgecicd.id
 }
 
 # Cost and usage
+module "dashboard_opencost" {
+  source = "./dashboards/opencost"
+
+  providers = {
+    signalfx = signalfx
+  }
+
+  tenant_names      = var.dashboard_variables.runner_k8s.tenant_names
+  dynamic_variables = var.dashboard_variables.runner_k8s.dynamic_variables
+  dashboard_group   = signalfx_dashboard_group.forgecicd.id
+}
+
 module "dashboard_billing" {
   source = "./dashboards/billing"